--- integration/tests/verification/happypath/main_test.go
@@ -1,12 +1,0 @@
-package happypath
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/verification"
-	"github.com/stretchr/testify/suite"
-)
-
-func TestHappyPath(t *testing.T) {
-	suite.Run(t, new(verification.VerificationTestSuite))
-}
