--- integration/tests/verification/systemchunk/main_test.go
@@ -1,12 +1,0 @@
-package systemchunk
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/verification"
-	"github.com/stretchr/testify/suite"
-)
-
-func TestVerifySystemChunk(t *testing.T) {
-	suite.Run(t, new(verification.VerifySystemChunkSuite))
-}
